cobalt/uapi, copperplate, alchemy, trank: make cobalt_event 32/64bit-neutral
Using long integers for representing event bit masks potentially
creates a portability issue for applications between 32 and 64bit CPU
architectures. This issue is solved by using 32bit integers on 32/64
bit machines, which is normally more than enough for encoding the set
of events received by a single event group.

In addition, this enables the sharing of the event value between a
64bit kernel support and 32bit libraries without thunks.

+++ include/cobalt/sys/cobalt.h
@@ -106,20 +106,20 @@ void cobalt_monitor_drain_all(cobalt_monitor_t *mon);
 int cobalt_monitor_drain_all_sync(cobalt_monitor_t *mon);
 
 int cobalt_event_init(cobalt_event_t *event,
-		      unsigned long value,
+		      unsigned int value,
 		      int flags);
 
 int cobalt_event_post(cobalt_event_t *event,
-		      unsigned long bits);
+		      unsigned int bits);
 
 int cobalt_event_wait(cobalt_event_t *event,
-		      unsigned long bits,
-		      unsigned long *bits_r,
+		      unsigned int bits,
+		      unsigned int *bits_r,
 		      int mode,
 		      const struct timespec *timeout);
 
 unsigned long cobalt_event_clear(cobalt_event_t *event,
-				 unsigned long bits);
+				 unsigned int bits);
 
 int cobalt_event_inquire(cobalt_event_t *event,
 			 struct cobalt_event_info *info,
